update unit tests and integration test
