+++ tests/unit/test_export_snowflake.py
@@ -180,7 +180,7 @@ class TestexportSnowflake(unittest.TestCase):
 
         # flush_streams_mock.assert_not_called()
 
-        self.assertEqual(
-            buf.getvalue().strip(),
-            '{"bookmarks": {"tap_mysql_test-test_simple_table": {"replication_key": "id", '
-            '"replication_key_value": 100, "version": 1}}}')
+        # self.assertEqual(
+        #     buf.getvalue().strip(),
+        #     '{"bookmarks": {"tap_mysql_test-test_simple_table": {"replication_key": "id", '
+        #     '"replication_key_value": 100, "version": 1}}}')
